feat: saving checkpoint getserversideprops refactor
